+++ HISTORY.md
@@ -1,4 +1,5 @@
 ## 2.7.3 (2015-04-18)
+* Fixed a potential memory violation in EvaluateVal() in hddata.f90
 * Bug fix in potlib.f90 and libutility.f90 that may cause memory leak
 
 ## 2.7.2 (2015-03-10)